requirements.txt
fastapi
uvicorn
python-multipart


sentence-transformers
faiss-cpu
numpy

pypdf
requests
